Use correct log levels for logging
A lot of the log messages are debug information, which is not commonly of much
value. I have changed these messages to use the 'debug' level. If a process does
have to be killed, I think this is a little more important, so I've left this
at to 'warn' level.